Book review and basket

+++ PustokLayout/PustokLayout/ViewModels/ReviewCreateViewModel.cs
@@ -5,7 +5,7 @@ namespace PustokLayout.ViewModels
     public class ReviewCreateViewModel
     {
         [Range(1, 5)]
-        public int Rate { get; set; }
+        public byte Rate { get; set; }
         [MaxLength(50)]
         public string Text { get; set; }
         public int BookId { get; set; }
